// EPS-Client/src/app/enums/statusType.enum.ts
export enum StatusType {
    LocationStatus = 1,
    AppointmentStatusPending = 2,
    AppointmentStatusConfirmed = 3,
    AppointmentStatusCancelled = 4,
    PersonStatusConfirmed = 5,
    PersonStatusCancelled = 6,
    AppointmentStatusRegistered = 7,
    PersonStatusRegistered = 8,
    ArticleStatus = 9,
    PersonStatusApplied = 10,
    InvoicingStatus = 11,
    Questionnaire = 12,
    AppointmentStatusPlanned = 13,
    TourplanImporter = 14,
    TradeDimensionImporter = 15
}