// EPS-Client/src/app/modules/api/model/appointmentResponse.ts
/**
 * Agency Pulse API v1.0
 * API für das Agency Pulse - Einsatzplanungssystem
 *
 * OpenAPI spec version: 1.0
 * Contact: development@de.cpm-int.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AppointmentStatusResponse } from './appointmentStatusResponse';
import { AttributeValueResponse } from './attributeValueResponse';
import { LocationResponse } from './locationResponse';
import { PersonResponse } from './personResponse';

export interface AppointmentResponse { 
    id?: string;
    associationId?: string;
    date?: Date;
    plannedFrom?: Date;
    plannedTo?: Date;
    timeFrom?: string;
    timeTo?: string;
    personToLocationAirlineDistance?: number;
    projectId?: string;
    assignedPerson?: PersonResponse;
    assignedLocation?: LocationResponse;
    appointmentStatus?: Array<AppointmentStatusResponse>;
    invoiceingAppointmentStatus?: Array<AppointmentStatusResponse>;
    categoryIds?: Array<string>;
    attributeValues?: Array<AttributeValueResponse>;
    parentAppointmentId?: string;
    hasChilds?: boolean;
    followUpType?: number;
    appStartTime?: Date;
    appEndTime?: Date;
}